+++ app/src/main/java/com/sopt/gongbaek/presentation/ui/groupdetail/navigation/GroupDetailNavGraph.kt
@@ -6,7 +6,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.sopt.gongbaek.presentation.model.NavigationRoute
 import com.sopt.gongbaek.presentation.ui.groupdetail.screen.GroupDetailRoute
-import com.sopt.gongbaek.presentation.ui.mygroup.navigation.navigateGroupRoom
+import com.sopt.gongbaek.presentation.ui.grouproom.navigation.navigateGroupRoomNavGraph
 
 fun NavGraphBuilder.groupDetailNavGraph(
     navController: NavHostController
@@ -19,7 +19,7 @@ fun NavGraphBuilder.groupDetailNavGraph(
             route = NavigationRoute.GroupDetailNavGraphRoute.GROUP_DETAIL
         ) {
             GroupDetailRoute(
-                navigateGroupRoom = { navController.navigateGroupRoom() }
+                navigateGroupRoom = { navController.navigateGroupRoomNavGraph() }
             )
         }
     }
