+++ app/src/main/java/com/sopt/gongbaek/presentation/ui/main/MainNavHost.kt
@@ -13,6 +13,7 @@ import com.sopt.gongbaek.presentation.model.NavigationRoute
 import com.sopt.gongbaek.presentation.ui.groupdetail.navigation.groupDetailNavGraph
 import com.sopt.gongbaek.presentation.ui.grouplist.navigation.groupListNavGraph
 import com.sopt.gongbaek.presentation.ui.groupregister.navigation.groupRegisterNavGraph
+import com.sopt.gongbaek.presentation.ui.grouproom.navigation.groupRoomNavGraph
 import com.sopt.gongbaek.presentation.ui.home.navigation.homeNavGraph
 import com.sopt.gongbaek.presentation.ui.mygroup.navigation.myGroupNavGraph
 import com.sopt.gongbaek.presentation.ui.onboarding.navigation.onboardingNavGraph
@@ -41,6 +42,7 @@ fun MainNavHost(
             groupRegisterNavGraph(navigator.navController)
             groupDetailNavGraph(navigator.navController)
             myGroupNavGraph(navigator.navController)
+            groupRoomNavGraph(navigator.navController)
             homeNavGraph(navigator.navController)
         }
     }
